resolve #724 non current mk appears in committee page if was replacement or chairperson
